feat(mobile + backend): More work on outings
Added initial backend code for getAllOuting function call.

Some part of it is hardcoded (see the TODOs), need to figure out the mapping of the VoteStepDTO

Rough mockup for the Flutter ViewAllOutings Page to verify correctness.

Will work on the ui for it when i wake up in roughly 6 hours